Make paging example a test.
